+++ static/src/javascripts/projects/common/modules/ui/bannerPicker.js
